Add RSS feed for the blog; load posts once.
