+++ package.json
@@ -36,6 +36,7 @@
 		"prettier-plugin-svelte": "^2.8.1",
 		"reading-time": "^1.5.0",
 		"rehype-picture": "^4.0.2",
+		"rss": "^1.2.2",
 		"sharp": "^0.31.2",
 		"svelte": "^3.55.0",
 		"svelte-check": "^2.9.2",
