+++ src/routes/+page.server.ts
@@ -1,0 +1,7 @@
+import type { PageLoad } from './$types';
+
+import { posts } from "$lib/server/posts";
+
+export const load: PageLoad = async () => {
+	return { postsTotal: posts.length, posts: posts.filter(post => post.featured) };
+};
